requirements.txt
# ============================================
# 📦 Python 패키지 목록
# ============================================
#
# 설치 방법: 터미널에서 아래 명령어 실행
#   pip install -r requirements.txt
#
# ============================================


# ── 1단계: TMDB 연동 ──

# requests: 인터넷으로 데이터를 가져오는 도구
# TMDB API에 "오늘 신작 알려줘" 요청을 보낼 때 사용
requests==2.32.3

# python-dotenv: .env 파일에서 API 키를 읽어오는 도구
# 코드에 키를 직접 넣지 않고 안전하게 불러올 수 있게 해줌
python-dotenv==1.1.0


# ── 2단계: Gemini 카피 생성 ──
# google-generativeai: Google Gemini AI API를 사용하는 도구
google-generativeai==0.8.5

# ── 3단계: 카드 이미지 생성 ──
# playwright: HTML 페이지를 이미지로 캡처하는 도구
playwright==1.51.0

# ── 5단계: 명언 수집 (역사 컨셉 확장) ──
# wikiquotes: Wikiquote에서 검증된 명언을 가져오는 도구
wikiquotes==3.1.0
